* Use cmake to compile project under mac.   * Fixed compiling errors related to cmake. * Use libspng instead of libpng. * Implement simplified MPlayerCore engine.   * Support CoreAudio output   * Support mp3 decoder with minimp3   * Support mp4 decoder with faad2   * Support flac decoder with flac

+++ LyricsLib/LyricsKeywordFilter.cpp
@@ -75,7 +75,7 @@ void CLyricsKeywordFilter::init(cstr_t fnFilter) {
                 // remove
                 utf16string strW;
 
-                utf8ToUCS2(node->strContent.c_str(), node->strContent.size(), strW);
+                utf8ToUCS2(node->strContent.c_str(), (int)node->strContent.size(), strW);
                 ucs2TableDelChars(strW.c_str());
             } else if (node->name == "Replace") {
                 // replace
